requirements.txt
python-telegram-bot==21.0
gspread==6.1.2
google-auth==2.32.0
python-dotenv==1.0.1
pytz==2024.1

# PostgreSQL dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
prometheus-client==0.19.0
